requirements.txt
google-generativeai==0.3.2
Pillow==10.2.0
streamlit==1.31.1
langchain-core==0.3.19
openai==1.54.4
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0
flask-cors
pdf2image==1.17.0